% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix_from_divisions.R
\name{adjacency_matrix_from_divisions}
\alias{adjacency_matrix_from_divisions}
\title{Generate an adjacency matrix based on how often MPs vote together}
\usage{
adjacency_matrix_from_divisions(divisions, chamber_id = "HoR",
  start_date = "2000-01-01", end_date = "2020-01-01")
}
\arguments{
\item{divisions}{the data frame containing the votes}

\item{chamber_id}{either 'HoR' for the House of Representatives or 'Senate' for the Senate}

\item{start_date}{the start of the time period to consider}

\item{end_date}{the end of the time period to consider}
}
\value{
a matrix where the $(i,j)$th entry is the number of times the ith legislator votes with the jth legislator in the period considered
}
\description{
Generate an adjacency matrix based on how often MPs vote together
}
